% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_build_contigency_table.R
\name{build_contingency_table}
\alias{build_contingency_table}
\title{Build contigensy table for binary variable}
\usage{
build_contingency_table(binary_variable_dataframe)
}
\description{
This function reads data frame of binary variable
and returns a contigency table for input into
chi-squared test
}
\examples{
build_contingency_table(binary_variable_dataframe)
}
