% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_check_in.R
\name{check_in}
\alias{check_in}
\title{Check the proportion of species that are annotated in the microbe-directory}
\usage{
check_in(microbe_directory = microbe_directory,
  species_list = species_list, guess_names = TRUE)
}
\arguments{
\item{microbe_directory}{species_list guess_names}
}
\description{
This function takes the microbe directory dataframe and the species list as input. It
reports the proportion of species in the input list that are present in the microbe directory.
If the parameter guess_names == TRUE then it will try and guess names of the input to match
the microbe directory. The function returns the species list (revised or original depending
if names were guessed)
}
\examples{
check_in(microbe_directory=microbe_directory, species_list=species_list, guess_names=TRUE)
}
