% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_format_identifiers.R
\name{format_identifiers}
\alias{format_identifiers}
\title{Format the taxonomic labels in the microbe directory data}
\usage{
format_identifiers(microbe_directory)
}
\description{
This function formats the taxonomic labels in the microbe directory
and adds them as rownames. It is from phylum -> species.
}
\examples{
format_identifiers(microbe_directory)
}
