% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_get_binary_variable.R
\name{get_binary_variable}
\alias{get_binary_variable}
\title{Get a binary variable from microbe directory}
\usage{
get_binary_variable(microbe_directory = microbe_directory,
  test_set = test_set, background_set = background_set,
  variable = "spore_forming")
}
\arguments{
\item{database}{test_set background_set variable}
}
\description{
This function returns a dataframe of one of the binary
variable from the microbe directory for each of the test_set
and background set
}
\examples{
get_binary_variable(microbe_directory=microbe_directory, test_set=test_set, background_set=background_set, variable="spore_forming")
}
