% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_get_quantitative_variable.R
\name{get_quantitative_variable}
\alias{get_quantitative_variable}
\title{Get a quantititative variable from microbe directory}
\usage{
get_quantitative_variable(microbe_directory = microbe_directory,
  test_set = test_set, background_set = background_set,
  variable = "optimal_ph")
}
\arguments{
\item{database}{test_set background_set}
}
\description{
This function returns a dataframe of one of the quantitative
variable from the microbe directory for each of the test_set
and background set
}
\examples{
get_quantitative_variable(microbe_directory=microbe_directory, test_set=test_set, background_set=background_set, variable="optimal_ph")
}
