% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_results.R
\name{get_results}
\alias{get_results}
\title{Get the results table}
\usage{
get_results(run_association_list)
}
\arguments{
\item{result}{from run_associations}
}
\description{
Return results table
}
\examples{
get_results(run_association_list)
}
