% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_plot_binary_variable.R
\name{plot_binary_variable}
\alias{plot_binary_variable}
\title{Plot binary variable in test and background sets}
\usage{
plot_binary_variable(contingency_table, label = "label")
}
\description{
Barplot of the distribution of values between test and
background sets of species
}
\examples{
plot_binary_variable(contingency_table)
}
