% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_plot_quantitative_variable.R
\name{plot_quantitative_variable}
\alias{plot_quantitative_variable}
\title{Plot quantitative variable in test and background sets}
\usage{
plot_quantitative_variable(quantitative_variable_dataframe,
  colors = c("grey", "black"))
}
\description{
Boxplot of the distribution of values between test and
background sets of species
}
\examples{
plot_quantitative_variable(quantitative_variable_dataframe)
}
