% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_results.R
\name{plot_results}
\alias{plot_results}
\title{Plot the results}
\usage{
plot_results(run_association_list)
}
\arguments{
\item{plot}{results from run_associations}
}
\description{
Plot the results that are stored after run_associations
}
\examples{
plot_results(run_association_list)
}
