% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_read_database.R
\name{read_database}
\alias{read_database}
\title{Read The Microbe Directory database}
\usage{
read_database(database)
}
\arguments{
\item{database}{in .csv format}
}
\description{
This function reads in the data in csv format.
}
\examples{
read_database("microbe-directory.csv")
}
