% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_read_list.R
\name{read_list}
\alias{read_list}
\title{Read lists for background or foreground sets of species}
\usage{
read_list(species_file)
}
\arguments{
\item{set}{of species (foreground or background)}
}
\description{
This function reads in the foreground/background sets (single column .tsv file)
}
\examples{
read_list("foreground.tsv"))
}
