% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_run_associations.R
\name{run_associations}
\alias{run_associations}
\title{Run the associations}
\usage{
run_associations(test_set_file, background_set_file, guess_names = TRUE)
}
\description{
This is the main function that performs the association
testing bewteen variables in The Microbe Directory and
species lists of interest.
}
\examples{
run_associations(test_set_file, background_set_file)
}
