% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_run_fisherexact.R
\name{run_chisquare}
\alias{run_chisquare}
\title{Chi-squared test on contingency table}
\usage{
run_chisquare(contingency_table, label = "label")
}
\arguments{
\item{contingency}{table}
}
\description{
Chi-squared test on binary variable counts
between test_set and background_set
}
\examples{
run_chisquare(contingency_table, label="label")
}
