% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_run_fisherexact.R
\name{run_fisherexact}
\alias{run_fisherexact}
\title{Fishers exact test}
\usage{
run_fisherexact(contingency_table, label = "label")
}
\arguments{
\item{contingency}{table}
}
\description{
Fisher exact test on binary variable counts
between test_set and background_set
}
\examples{
run_fisherexact(contingency_table, label="label")
}
