% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mdat_run_wilcoxtest.R
\name{run_wilcoxtest}
\alias{run_wilcoxtest}
\title{Wilcoxon rank sum test on quantitative variables}
\usage{
run_wilcoxtest(quantitative_variable_dataframe, label = "label")
}
\description{
Wilcoxon test on quantitative variables
between test_set and background_set
}
\examples{
run_wilcoxtest(quantitative_variable_dataframe, label="label")
}
